Moving LICENSE and README files back into base of project.
These might get in the way of someone trying to import the project
directly but this is realistically where they belong and they're
not hard to get rid of should someone desire it.

At the very least, GitHub seems to want them in the base of the project
which is probably reason enough.